Section templates
What your workflow is like.

I built out skeletons and integrated all Liquid and Liquid logic. Now I will go ahead and style/populate the sections.

What was challenging / what wasn’t? (Please be honest here. If something is challenging, that’s ok and it will just help us see where our strengths and weaknesses might fit with yours and vice versa.)

Nothing was really challenging. I assessed the Figma and added in the necessary fields required to populate the sections.

Any general decisions you make along the way.

I am breaking all the steps down into how they need to be approached and developed. I am doing each section one by one, and making sure, in the theme editor, everything is populating as expected.

What would you discuss with a designer through the process?

Do the titles require any options for different font sizing options in the section, or are they fixed?

What are a few code decisions you made and why?

I noticed the text and columns will need to be blocks. With that said there are a list of links to collections in the left column, so to make it easier for the editor, I utilized the collection list option so the client can select multiple collections and  links and and collection titles will populate. I limited it to only 4 links.