Section updates and styles
What your workflow is like.

I added in a few styles for the sections, and also altered a few of the sections and how they're structured and allow for different content and consideration for responsiveness. I will then add in content in the theme editor, make a pull from the theme, and then tweak the sections and style and update the structure of the header following this.

What was challenging / what wasn’t? (Please be honest here. If something is challenging, that’s ok and it will just help us see where our strengths and weaknesses might fit with yours and vice versa.)

I decided, for the image and columns section, to utilize Shopify's new blocks system. I took a bit of time to refresh myself to re-learn this, but I since the blocks do ask for a single link and a collection list, for the sets of links, I thought it was a great opportunity. I used private blocks so they're only designated to the column block.

Any general decisions you make along the way.

I updated the image and columns sections to accommodate Shopify's new blocks system.

What would you discuss with a designer through the process?

I would say, the image quality needs to be better. The images are coming in a little more pixelated than I would like.

What are a few code decisions you made and why?

I added in some conditions in the hero section I built to allow for the user to update the height of the image. I also added a desktop and mobile option, as the images are a little different in the design from each other. I also did something similar for the image and text to allow for different image heights for desktop and mobile screens, as I noticed the image was much shorter on mobile, compared to desktop. Essentially giving the user more flexibility to make more style fixes. I also added the user the option to change the background colour of the icons and columns section, and also allowed the option for a head, just in case.